# maestro-hive/sunday_com_backup_20251005/backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sunday.com API
  description: |
    Comprehensive API documentation for Sunday.com work management platform.

    ## Authentication
    All API endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for unauthenticated users

    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "error": {
        "type": "validation_error",
        "message": "Invalid request data",
        "details": {},
        "requestId": "uuid"
      }
    }
    ```

    ## Pagination
    List endpoints support cursor and offset-based pagination:
    ```
    GET /api/v1/boards?page=1&limit=20
    ```

  version: 1.0.0
  contact:
    name: Sunday.com Team
    email: api@sunday.com
    url: https://sunday.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sunday.com/terms
servers:
  - url: https://api.sunday.com/api/v1
    description: Production server
  - url: https://staging-api.sunday.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []

# ============================================================================
# AUTHENTICATION
# ============================================================================
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      description: Get all workspaces in the user's organizations
      parameters:
        - in: query
          name: organizationId
          schema:
            type: string
            format: uuid
          description: Filter by organization ID
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Workspaces retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Workspaces]
      summary: Create workspace
      description: Create a new workspace in an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Workspace'
        '422':
          $ref: '#/components/responses/ValidationError'
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      description: Get workspace details by ID
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Workspaces]
      summary: Update workspace
      description: Update workspace details
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Workspace updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Workspaces]
      summary: Delete workspace
      description: Soft delete a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '204':
          description: Workspace deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'

# ============================================================================
# BOARDS
# ============================================================================
  /boards/workspace/{workspaceId}:
    get:
      tags: [Boards]
      summary: List workspace boards
      description: Get all boards in a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - in: query
          name: folderId
          schema:
            type: string
            format: uuid
          description: Filter by folder ID
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Boards retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Board'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workspace ID
...
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        settings:
          type: object
          additionalProperties: true
        isPrivate:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Board:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          type: object
          additionalProperties: true
        isPrivate:
          type: boolean
        position:
          type: number
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

    # Request Schemas
...
    CreateWorkspaceRequest:
      type: object
      required: [name, organizationId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        organizationId:
          type: string
          format: uuid
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: "#6B7280"
        isPrivate:
          type: boolean
          default: false
        settings:
          type: object
          additionalProperties: true
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        isPrivate:
          type: boolean
        settings:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "not_found_error"
              message: "Resource not found"
              requestId: "req_123456"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "validation_error"
              message: "Invalid request data"
              details:
                name: ["Name is required"]
                email: ["Invalid email format"]
              requestId: "req_123456"
...